scheduling, current info and monthly bills details
