+++ src/main/java/com/snezana/introtelecom/exceptions/RestAPIErrorMessage.java
@@ -9,6 +9,7 @@ public enum RestAPIErrorMessage {
     INVALID_DATE_TIME_FORMAT,
     INVALID_STARTDATETIME_OR_ENDDATETIME,
     INVALID_STARTDATE_OR_ENDDATE,
+    INVALID_DATETIME_INTERVAL,
     INVALID_INPUT_FORMAT,
     VALUE_CANNOT_BE_NEGATIVE,
     PARAMETER_CANNOT_BE_NULL;
