+++ src/main/java/com/snezana/introtelecom/mapper/PackageFrameMapper.java
@@ -19,16 +19,6 @@ import java.util.List;
 public abstract class PackageFrameMapper {
     Logger log = org.slf4j.LoggerFactory.getLogger(PackageFrameMapper.class);
 
-    public abstract PackageFrame packageFrameSaveDtoToPackageFrame(PackageFrameSaveDTO packageFrameSaveDto, @Context PhoneRepo phoneRepo);
-
-    @AfterMapping
-    protected void afterPackageFrameSaveDtoToPackageFrame(PackageFrameSaveDTO packageFrameSaveDto, @MappingTarget PackageFrame packageFrame, @Context PhoneRepo phoneRepo) {
-        Phone phone = phoneRepo.findByPhoneNumber(packageFrameSaveDto.getPhoneNumber());
-        packageFrame.setPhone(phone);
-        packageFrame.setPackfrStatus(StatusType.PRESENT.getStatus());
-        setPackageFrameByPackageCode(packageFrame);
-    }
-
     @Mapping(source = "packageFrame", target = "packfrCls", qualifiedByName = "calls")
     @Mapping(source = "packageFrame", target = "packfrSms", qualifiedByName = "messages")
     @Mapping(source = "packageFrame", target = "packfrInt", qualifiedByName = "internet")
@@ -91,42 +81,6 @@ public abstract class PackageFrameMapper {
     protected void beforePackageFrameToPackageFrameViewDTO(PackageFrame packageFrame, @MappingTarget PackageFrameViewDTO packageFrameViewDTO){
         packageFrameViewDTO.setPhoneNumber(packageFrame.getPhone().getPhoneNumber());
         packageFrameViewDTO.setPackageCode(packageFrame.getPhone().getPackagePlan().getPackageCode());
-    }
-
-    private void setMonthlyPackageFrameQuota (PackageFrame packageFrame, int monthlyQuotaCls, int monthlyQuotaSms, String monthlyQuotaInt, String monthlyQuotaAsm, String monthlyQuotaIcl, String monthlyQuotaRmg){
-        packageFrame.setPackfrCls(monthlyQuotaCls);
-        packageFrame.setPackfrSms(monthlyQuotaSms);
-        packageFrame.setPackfrInt(new BigDecimal(monthlyQuotaInt));
-        packageFrame.setPackfrAsm(new BigDecimal(monthlyQuotaAsm));
-        packageFrame.setPackfrIcl(new BigDecimal(monthlyQuotaIcl));
-        packageFrame.setPackfrRmg(new BigDecimal(monthlyQuotaRmg));
-
-    }
-
-    private void setPackageFrameByPackageCode(PackageFrame packageFrame) {
-        String packageCode = packageFrame.getPhone().getPackagePlan().getPackageCode();
-        PackagePlanType packagePlanType = PackagePlanType.findByKey(packageCode);
-        switch (packagePlanType) {
-            case PRP01:
-                setMonthlyPackageFrameQuota(packageFrame, 200, 200, "0.00", "0.00", "0.00", "0.00");
-                break;
-            case PRP02:
-                setMonthlyPackageFrameQuota(packageFrame, 200, 200, "10000.00", "0.00", "0.00", "0.00");
-                break;
-            case PST11:
-                setMonthlyPackageFrameQuota(packageFrame, 300, 300, "10000.00", "0.00", "200.00", "200.00");
-                break;
-            case PST12:
-                setMonthlyPackageFrameQuota(packageFrame, 400, 400, "10000.00", "5000.00", "200.00", "200.00");
-                break;
-            case PST13:
-                setMonthlyPackageFrameQuota(packageFrame, -1, -1, "15000.00", "5000.00", "200.00", "200.00");
-                break;
-            case PST14:
-                setMonthlyPackageFrameQuota(packageFrame, -1, -1, "-1.00", "10000.00", "200.00", "200.00");
-                break;
-            default:
-        }
     }
 
 }
